refactor: move navigation to components
