+++ apps/broccoli/src/components/ApplicationLayout/ApplicationLayout.tsx
@@ -1,7 +1,7 @@
 import { useAuth } from '@clerk/clerk-react';
 import { useEffect } from 'react';
 import { Outlet, useNavigate } from 'react-router-dom';
-import { Navigation } from '../../features/Navigation';
+import { Navigation } from '../../components/Navigation';
 
 export const ApplicationLayout = () => {
   const { isLoaded, userId } = useAuth();
