--- apps/broccoli/src/features/Navigation/components/NavLinks/index.ts
@@ -1,1 +1,0 @@
-export { NavLinks } from './NavLinks';
